made the app successfully restart
