+++ src/renderer/hooks/index.ts
@@ -4,6 +4,7 @@ import useCrawlSockets from './useCrawlSockets';
 import useCleanSockets from './useCleanSockets';
 import useReadIpc from './ipc/useReadIpc';
 import useWriteIpc from './ipc/useWriteIpc';
+import {useIpcEffect} from './ipc/utils';
 import useEventListener from './useEventListener';
 import useFormContext from './useFormContext';
 import useFormSelect from './useFormSelect';
@@ -20,6 +21,7 @@ export {
   useBooleanState,
   useCrawlSockets,
   useCleanSockets,
+  useIpcEffect,
   useReadIpc,
   useWriteIpc,
   useEventListener,
