+++ src/renderer/hooks/ipc/useWriteIpc.tsx
@@ -33,7 +33,7 @@ function useWriteIpc({
       buttonLabel: downloadOptions.buttonLabel,
       defaultPath: downloadOptions.defaultPath,
       filters: [
-        {extensions: ['txt'], name: 'txt'},
+        {extensions: ['json'], name: 'json'},
         {extensions: ['*'], name: 'All Files'},
       ],
       title: downloadOptions.title,
